% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_medium_zoom1}
\alias{case_medium_zoom1}
\title{Medium Challenge Changepoint Timeseries DataCube}
\format{
An object of class \code{stars} of dimension 62 x 63 x 11.
}
\usage{
case_medium_zoom1
}
\description{
Timeseries datacube. Each pixel is a timeseries, some pixels
have a downwards jump after some timepoint. Such timepoint
is called a changepoint.
}
\seealso{
\link[tjdc]{case_medium_zoom1_details} for true changepoint locations and more.
}
\keyword{datasets}
