% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_medium_zoom1_details}
\alias{case_medium_zoom1_details}
\title{Medium Challenge Changepoint Timeseries DataCube}
\format{
An object of class \code{list} of length 4.
}
\usage{
case_medium_zoom1_details
}
\description{
Details for the \link[tjdc]{case_medium_zoom1} datacube.
}
\details{
- Table with pixelwise regression parameters in the simulation, jump time (or not),
etc.

- bounding box info

- time-variable values
}
\keyword{datasets}
