% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_m0.3.R
\name{predict_m0.3}
\alias{predict_m0.3}
\title{Prediction the series per cell}
\usage{
predict_m0.3(x, cells, ..., s)
}
\arguments{
\item{x}{object fitted by this package's tj_fit_m0.3}

\item{cells}{which cells to predict? Default: all}

\item{s}{output of summarise_m0.3}
}
\description{
Prediction the series per cell
}
