% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_neighbours.R
\name{tj_cell_neighbours}
\alias{tj_cell_neighbours}
\title{Raster Cell Neighbourhoods}
\usage{
tj_cell_neighbours(i, nr, nc, ...)
}
\arguments{
\item{i}{focal cell}

\item{nr}{number of rows in the xy-grid}

\item{nc}{number of columns in the xy-grid}

\item{...}{ignored}
}
\description{
Raster Cell Neighbourhoods
}
\details{
Non-toroidal, 8 nearest aka queen neighbours.

Will assume cell-numbering
cell = col + (row-1) * ncol.
}
