% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_neighbours.R
\name{tj_cell_neighbours_old}
\alias{tj_cell_neighbours_old}
\title{Raster Cell Neighbourhoods}
\usage{
tj_cell_neighbours_old(i, nr, nc, ...)
}
\arguments{
\item{i}{focal cell}

\item{nr}{number of rows in the xy-grid}

\item{nc}{number of columns in the xy-grid}

\item{...}{ignored}
}
\value{
A vector of indices. If `type='sqexp'`, will return the weights in the attribute 'weight'.
}
\description{
Raster Cell Neighbourhoods
}
\details{
Non-toroidal neighborhood, so at edges there will be less neighbours.

Will assume cell-numbering
cell = col + (row-1) * ncol.
}
