% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gaussian_regression_with_mcmc-truncated-jump-v6.1.R,
%   R/gaussian_regression_with_mcmc-truncated-jump-v6.R
\name{tj_cfg_m0.6}
\alias{tj_cfg_m0.6}
\title{Compile paramaters for m0.6}
\usage{
tj_cfg_m0.6(...)

tj_cfg_m0.6(...)
}
\description{
Compile paramaters for m0.6

Compile paramaters for m0.6
}
