% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.5.R
\name{tj_divide_bbox}
\alias{tj_divide_bbox}
\title{Divide bounding box into subsets, return polygons}
\usage{
tj_divide_bbox(bb, power_of_2 = 2, q = power_of_2, expand = 0)
}
\description{
Divide both sides by a factor 2^q
}
