% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.5.R
\name{tj_fit_m0.3_dac}
\alias{tj_fit_m0.3_dac}
\title{Wrapper to fit model 3 to a mosaic}
\usage{
tj_fit_m0.3_dac(
  x,
  timevar = "z",
  attrvar = "values",
  dat = NULL,
  prior_theta = list(m = c(a = 0, b = 0, d = 0), S = diag(c(1, 1, 1) * 1e+05)),
  prior_sigma2 = c(shape = 2, rate = 1),
  prior_k = 0.5,
  gamma = 0,
  verbose = TRUE,
  dbg = FALSE,
  method = "mc",
  niter = 100,
  subsets = NULL,
  cells_to_ignore = NULL,
  truncate_jump_at_mean = 0,
  ...,
  ctrl = list(eps = 0.1, burnin = 0.5, thin_steps = 1),
  stitcher = tj_stitcher_m0.5_v1.1,
  ncores = 1,
  fitpath = NULL,
  fitid = "youridhere",
  just_fit = !is.null(fitpath),
  keep_all_in_memory = TRUE,
  recalc = FALSE,
  keep_hist = FALSE
)
}
\arguments{
\item{x}{stars datacube}

\item{subsets}{list of cell indices}

\item{...}{ignored}
}
\description{
Split datacube into sub-rasters, fit model 0.3's, and stitch together. Divide-and-conquer.
}
\details{
Wrap fitting of m0.3 in subsets for multicore computation.
}
