% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_regression_with_mcmc-v3.R
\name{tj_fit_m0.3s}
\alias{tj_fit_m0.3s}
\title{Fit m0.3 alternative parametrisation}
\usage{
tj_fit_m0.3s(..., cfg = tj_cfg_m0.3())
}
\description{
Wrapper for m0.3 to be called with data and a single list-item of run-parameters
}
\seealso{
[tj_cfg_m0.3] for creating required list.
}
