% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.5.R
\name{tj_fit_m0.5}
\alias{tj_fit_m0.5}
\title{Wrapper to fit model 3 to a mosaic}
\usage{
tj_fit_m0.5(
  x,
  timevar = "year",
  attrvar = "values",
  tiling,
  cfg,
  ...,
  ncores = 1,
  verbose = TRUE,
  verbose2 = verbose,
  fitpath = NULL,
  fitid = "youridhere",
  keep_all_in_memory = TRUE,
  recalc = FALSE
)
}
\arguments{
\item{x}{stars datacube}

\item{tiling}{given by [tj_divide_raster()]}

\item{cfg}{m0.3 parameters, see [tj_cfg_m0.3()]}

\item{...}{parameters passed on to [tj_fit_m0.3()] (in case cfg not used)}

\item{cores}{number of cores to use in multicore computation using `doParallel`}
}
\value{
List of fits (keep_all_in_memory=TRUE) or list of files where the fits are stored (fitpath provided).
Note: The fits have cell_info element where the 'truecell' column provides the mapping to the original 'x'.
}
\description{
(OBSOLETE) Split datacube into sub-rasters, fit model 0.3's, and stitch together. Divide-and-conquer.
}
\details{
Wrap fitting of m0.3 in subsets for multicore computation. The tiling splits the stars-object
into subrasters, then [tj_fit_m0.3()] is called on each subraster.

This has been replaced by the more general tj_fit_m0.x.
}
\seealso{
[tj_cfg_m0.3()] [tj_fit_m0.x]
}
