% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gaussian_regression_with_mcmc-truncated-jump-v6.1.R,
%   R/gaussian_regression_with_mcmc-truncated-jump-v6.R
\name{tj_fit_m0.6s}
\alias{tj_fit_m0.6s}
\title{Fit m0.6 alternative parametrisation}
\usage{
tj_fit_m0.6s(..., cfg = tj_cfg_m0.6())

tj_fit_m0.6s(..., cfg = tj_cfg_m0.6())
}
\description{
Wrapper for m0.6 to be called with data and a single list-item of run-parameters

Wrapper for m0.6 to be called with data and a single list-item of run-parameters
}
\seealso{
[tj_cfg_m0.6] for creating required list.

[tj_cfg_m0.6] for creating required list.
}
