% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.x_wrapper.R
\name{tj_fit_m0.x}
\alias{tj_fit_m0.x}
\title{Wrapper to fit models of the 0.x-family to a mosaic}
\usage{
tj_fit_m0.x(
  x,
  timevar = "year",
  attrvar = "values",
  tiling,
  cfg,
  ...,
  ncores = 1,
  verbose = TRUE,
  verbose2 = verbose,
  fitpath = NULL,
  fitid = "youridhere",
  keep_all_in_memory = TRUE,
  model_variant = "m0.3",
  recalc = FALSE
)
}
\arguments{
\item{x}{stars datacube}

\item{tiling}{given by [tj_divide_raster()]}

\item{cfg}{m0.x parameters, see [tj_cfg_m0.3(), tj_cfg_m0.6()]}

\item{...}{parameters passed on to [tj_fit_m0.3(), tj_fit_m0.6()] (in case cfg not used)}

\item{cores}{number of cores to use in multicore computation using `doParallel`}
}
\value{
List of fits (keep_all_in_memory=TRUE) or list of files where the fits are stored (fitpath provided).
Note: The fits have cell_info element where the 'truecell' column provides the mapping to the original 'x'.
}
\description{
Split datacube into sub-rasters, fit model 0.x's, and stitch together. Divide-and-conquer.
}
\details{
Wrap fitting of m0.x in subsets for multicore computation. The tiling splits the stars-object
into subrasters, then [tj_fit_m0.3(), tj_fit_m0.6()] is called on each subraster.
}
\seealso{
[tj_cfg_m0.3(), tj_cfg_m0.6()]
}
