% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall.R
\name{tj_mann_kendall}
\alias{tj_mann_kendall}
\title{Mann-Kendall Trend Test}
\usage{
tj_mann_kendall(x, ..., est_beta = TRUE, calc_slope = est_beta, check = TRUE)
}
\arguments{
\item{x}{Time series, either a vector of values or n x 2 matrix of value,time pairs}

\item{...}{ignored}

\item{est_beta}{calc_slope (legacy)}

\item{calc_slope}{Estimate linear coefficient as well?}

\item{check}{check the data. If you give nx2 matrix then this FALSE saves time.}
}
\description{
1D, just the regular old.
}
\details{
No ties correction. This exists in the package mainly for sanity checks.
}
