% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_calc_wrapper_obsolete.R
\name{tj_split_calc_wrapper}
\alias{tj_split_calc_wrapper}
\title{Split-Calc-Merge Wrapper}
\usage{
tj_split_calc_wrapper(
  x,
  nx,
  ny,
  buffer = c(1, 1),
  fun,
  ...,
  dbg = FALSE,
  byrowcol = TRUE
)
}
\arguments{
\item{x}{stars-object}

\item{nx}{Number of x-dimension divisions}

\item{ny}{Number of y-dimension divisions}

\item{buffer}{Extent each sub-cube* by this many cells (byrowcol=TRUE) or spatial units (byrowcol=FALSE)}

\item{fun}{The function to call per piece. Should return either raster or cube of identical dimensions as input for it.}

\item{...}{passed to fun in argument list after the piece}

\item{dbg}{Print messages?}

\item{byrowcol}{Split by row-col indices instead of spatial extents? Default is TRUE}
}
\value{
The output is a list. The first element of this list, called 'result', should be identical to the output of simply calling 'fun(x, ...)'. The remaining elements contain various information related to the calculation, e.g. time spent.
}
\description{
Takes a datacube, splits it into spatial pieces, applies a function on each sub-cube, and then merges the results together.
}
\details{
This function uses the 'future' package for multi-core processing, just define a [future::plan()] before calling it.

The 'fun' should take a rasterStack as input and also should return a rasterStack
(or raster).
}
\note{
If byrowcol=FALSE: a stitching message about different buffers might appear (from 'mergeRaster()'), the (nx,ny) pair
is problematic for 'mergeRaster' and the resulting raster will have artifacts in the splitted edges. Different (nx,ny) is the only way to get around this at the moment.
}
\references{
Neeti, N. and Eastman J.R. (2011) A Contextual Mann-Kendall Approach for the Assesment of Trend Significance in Image Time Series, \emph{Transactions in GIS}
}
