% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_polygons.R
\name{tj_stars_to_polygon_union}
\alias{tj_stars_to_polygon_union}
\title{Raster to Polygons
Take stars-raster and create a polygons of those pixels with >0 values using st_union}
\usage{
tj_stars_to_polygon_union(r, which = 1)
}
\arguments{
\item{r}{stars object}
}
\value{
sf-object of polygons
}
\description{
Raster to Polygons
Take stars-raster and create a polygons of those pixels with >0 values using st_union
}
