% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.5.R
\name{tj_stitcher_m0.5_v1.2}
\alias{tj_stitcher_m0.5_v1.2}
\title{Stitch a mosaic model fit}
\usage{
tj_stitcher_m0.5_v1.2(list_of_fits, subsets)
}
\description{
Average overlapping regions' pixel estimates. Version 2, use included cell mappings.
}
