% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_m0.x_wrapper.R
\name{tj_stitcher_m0.x}
\alias{tj_stitcher_m0.x}
\title{Stitch a mosaic model fit}
\usage{
tj_stitcher_m0.x(list_of_fits, tiling)
}
\description{
Average overlapping regions' pixel estimates. Old format.
}
