% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_m0.5.R
\name{tj_summarise_m0.5}
\alias{tj_summarise_m0.5}
\title{Summarise fit of M0.5 stiched results}
\usage{
tj_summarise_m0.5(x, ..., cells)
}
\arguments{
\item{x}{object fitted by this package's tj_fit_m0.5 and then stitched}
}
\description{
Summarise fit of M0.5 stiched results
}
