% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_m0.5.R
\name{tj_trace_m0.3_dac}
\alias{tj_trace_m0.3_dac}
\title{Retrieve trace for particular cell from list of mosaic fits}
\usage{
tj_trace_m0.3_dac(x, cell, add_sub_info = FALSE)
}
\arguments{
\item{x}{list of fits from the tj_fit_m0.3_dac}

\item{cell}{which cell, in the original raster}
}
\description{
Retrieve trace for particular cell from list of mosaic fits
}
