% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_m0.3.R
\name{trace_m0.3}
\alias{trace_m0.3}
\title{Get the history}
\usage{
trace_m0.3(x, cell, ...)
}
\arguments{
\item{x}{fit_m0.3 with `keep_hist=TRUE`}

\item{cell}{which cell}
}
\description{
Compile the MCMC trace for one cell
}
\details{
Combine with posterior-package.
}
