% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_neighbours.R
\name{tj_cell_neighbours}
\alias{tj_cell_neighbours}
\title{Raster Cell Neighbourhoods}
\usage{
tj_cell_neighbours(i, nr, nc)
}
\arguments{
\item{i}{focal cell}

\item{nr}{number of rows in the xy-raster}

\item{nc}{number of columns in the xy-raster}
}
\description{
Raster Cell Neighbourhoods
}
\details{
Non-toroidal.

NOTE we assume that the cell indices go first row then col, i.e. cell=1,2,3... are (row,col) (1,1), (1,2), (1,3)...
AND  we assume that rows go in "bottom to top" fashion in persieved y-dimension. Make sure data is like this! Stars::as_tibble
might have negative dimension deltas.
}
