% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input.R
\name{tj_stars_to_data}
\alias{tj_stars_to_data}
\title{Process input stars to legacy data format}
\usage{
tj_stars_to_data(x, timevar = "z", attrvar = "values")
}
\arguments{
\item{x}{stars object}
}
\description{
Process input stars to legacy data format
}
